Fixed a problem of inclusion of stdlib.h
A method in the class Pin uses the method std::atoi which is found in std::basic_string. This class is available including "stdlib.h". Using both g++ (6.3.1) and cmake with the c++11 standard it works ok, but @FightingForFun reported an error in Ubuntu 16.04 saying atoi is not defined.
I've now included stdlib.h in that class so it should not give that error anymore.